Drop main search
Doesn't work now that the main docs is on furo, but also the flyout option for subprojects also correctly searches sibling projects so it's no longer necessary.

Instead point people to the addon search in the flyout.
Hopefully at some point the addon search will be used natively for the Furo search, making this note redundant.